Restore GRPC methods in the main package for older versions of containerd
